Conflicts:
	.gitignore
	BizHawk.Client.Common/PathManager.cs
	BizHawk.Client.Common/lua/LuaSandbox.cs
	BizHawk.Client.EmuHawk/AVOut/FFmpegWriter.cs
	BizHawk.Client.EmuHawk/AVOut/SynclessRecordingTools.cs
	BizHawk.Client.EmuHawk/CustomControls/VirtualListView.cs
	BizHawk.Client.EmuHawk/DisplayManager/DisplayManager.cs
	BizHawk.Client.EmuHawk/GLManager.cs
	BizHawk.Client.EmuHawk/Program.cs
	BizHawk.Client.EmuHawk/movie/EditSubtitlesForm.cs
	BizHawk.Client.EmuHawk/movie/PlayMovie.cs
	BizHawk.Emulation.Common/Sound/Utilities/SpeexResampler.cs
	BizHawk.Emulation.Cores/Consoles/Nintendo/SNES/LibsnesApi.cs
	BizHawk.Emulation.Cores/Consoles/Nintendo/SNES/LibsnesApi_CMD.cs
	BizHawk.Emulation.Cores/Consoles/Nintendo/SNES/LibsnesCore.cs
	BizHawk.Emulation.Cores/Consoles/Sony/PSX/Octoshock.ITraceable.cs
	BizHawk.Emulation.Cores/LibretroOld/LibRetro.cs
	Version/VersionInfo.cs
	libsnes/bsnes/target-libsnes/libsnes_pwrap.cpp

+++ BizHawk.Client.EmuHawk/CustomControls/Win32.cs
@@ -311,9 +311,9 @@ namespace BizHawk.Client.EmuHawk
 			public int dwQuality;
 			public int dwBytesPerSecond;
 			public int dwFlags;
-			public int lpFormat;
+			public IntPtr lpFormat;
 			public int cbFormat;
-			public int lpParms;
+			public IntPtr lpParms;
 			public int cbParms;
 			public int dwInterleaveEvery;
 		}
@@ -472,7 +472,7 @@ namespace BizHawk.Client.EmuHawk
 		}
 
 		[DllImport("user32.dll", CharSet = CharSet.Auto)]
-		public static extern int SendMessage(IntPtr hWnd, int msg, int wParam, int lParam);
+		public static extern IntPtr SendMessage(IntPtr hWnd, int msg, IntPtr wParam, IntPtr lParam);
 
 		[DllImport("Kernel32.dll", EntryPoint = "RtlZeroMemory", SetLastError = false)]
 		public static extern void ZeroMemory(IntPtr dest, uint size);
